[6/x] change venft enable logic in project tools drawer
