+++ src/constants/featureFlags.ts
@@ -2,4 +2,5 @@ export const FEATURE_FLAGS = {
   V1_TOKEN_SWAP: 'v1TokenSwap',
   NFT_REWARDS: 'nftRewards',
   VENFT: 'veNft',
+  VENFT_CREATOR: 'veNftCreator',
 }
